Refactor: move weekly summary creation function to useCreateSummary hook
