+++ src/utils/timeCalculator.ts
